+++ lib/simple_form/action_view_extensions/form_helper.rb
@@ -16,7 +16,11 @@ module SimpleForm
         unless options[:html].key?(:novalidate)
           options[:html][:novalidate] = !SimpleForm.browser_validations
         end
-        options[:html][:class] = [SimpleForm.form_class, simple_form_css_class(record, options)].compact.join(" ")
+        if options[:html].key?(:class)
+          options[:html][:class] = [SimpleForm.form_class, options[:html][:class]].compact
+        else
+          options[:html][:class] = [SimpleForm.form_class, SimpleForm.default_form_class, simple_form_css_class(record, options)].compact
+        end
 
         with_simple_form_field_error_proc do
           form_for(record, options, &block)